Make pkg parameter to Wsdl case class optional
The package parameter is not necessary for the wsdl2java cli tool, and in some cases it is harmful to have to specify it, especially if the WSDL is composed of many schemas which may have overlapping elements. In this case it is better to omit the package parameter, this emits the classes into packages according their schema names. That ensures that there are no conflicts.